requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
blis==1.3.0
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
cryptography==44.0.3
cymem==2.0.11
distlib==0.3.9
Django==5.2.1
django-tailwind==4.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl  
filelock==3.18.0
idna==3.10
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
packaging==25.0
pandas==2.2.3
passlib==1.7.4
phonenumbers==8.13.55
pipenv==2025.0.3
platformdirs==4.3.8
preshed==3.0.10
presidio_analyzer==2.2.358
presidio_anonymizer==2.2.358
presidio_structured==0.0.5a0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.9.0
pymongo==4.13.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==14.0.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
tailwind==3.1.5b0
thinc==8.3.6
tldextract==5.3.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.31.2
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
numpy>=1.24.0  
scikit-learn>=1.4.0
python-decouple==3.8  
aiohttp==3.8.5
matplotlib==3.8.4
seaborn==0.13.2
google-generativeai==0.3.2  
playwright==1.40.0  
pytest==7.4.3  
pytest-django==4.7.0
pyapacheatlas==0.16.0  
kafka-python==2.0.2  

